fix XY

+++ main/SoulmateMatrix.h
@@ -129,7 +129,7 @@ int16_t gridIndex(int16_t x, int16_t y) {
   }
 }
 
-int16_t gridIndexHorizontal(int16_t x, int16_t y) {
+int16_t gridIndexHorizontal(uint8_t x, uint8_t y) {
   if (y > LED_ROWS) return -1;
   if (x > LED_COLS) return -1;
   if (x < 0) return -1;
@@ -152,7 +152,7 @@ int16_t gridIndexHorizontal(int16_t x, int16_t y) {
   return -1;
 }
 
-uint16_t XY(uint16_t x, uint16_t y) {
+uint16_t XY(uint8_t x, uint8_t y) {
   return gridIndexHorizontal(x, y);
 }
